Move docker terminal integration test to own file
